backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pyyaml>=6.0.0
openai>=1.0.0
python-dotenv>=1.0.0
